Add the initial documentation and spec for DiffX.
This goes over the rationale for why we need a new diff file format,
answers questions about the whys and hows, and includes a specification
for the format itself.

The spec is not complete, and some information like the metadata format
and more comprehensive examples need to be worked on still

This does provide the beginnings of a `diffx` Python package as well,
which currently just contains the beginnings of a Pygments lexer used by the
docs.