Update the sections on metadata for the new, upcoming grammar.
We didn't have a grammar defined for metadata sections in the past, and
while we've loosely modeled our metadata structure off of YAML and
Git's metadata format, neither option was really sufficient.

We'll be introducing a grammar for metadata. That's not here in this
change, but in preparation, all sample metadata in the spec has been
updated for the planned grammar. The section on the metadata language
and the reserved future `format` option for the metadata sections has
been updated accordingly.

Testing Done:
Built the docs. Checked the relevant sections to make sure they were
using the right syntax.

Reviewed at https://reviews.reviewboard.org/r/11690/