Rework the information on section lengths.
This clarifies the `length=` option for sections. It explains when a
length is required and when it isn't, how a length is calculated, and
the rationale for the lengths.

Lengths are no longer optional. Content sections (preambles, metadata,
diffs) now require it.

Some lengths in examples have been fixed to be accurate for the
provided content.

Testing Done:
Built the docs and read through the affected sections.

Tested all lengths in the examples to make sure they were now correct.

Reviewed at https://reviews.reviewboard.org/r/11691/