Fix generating the docs on Python 3.
The Sphinx `conf.py` was generated in a Python 2-only mode, with `u`
prefixes for strings. This changes it to use native strings, and to
include a `unicode_literals` import at the top.

Testing Done:
Successfully built the docs using Sphinx on Python 2 and 3.

Reviewed at https://reviews.reviewboard.org/r/11695/