Add and flesh out examples throughout the spec.
This adds examples for all options and metadata keys, and both improves
and fixes up some of the examples for sections.

There's still room for improvement here, but it should make it a lot
easier for others to ensure that their implementations are correct.

Testing Done:
Built the docs. Checked all the examples to ensure they were correct.

Reviewed at https://reviews.reviewboard.org/r/11698/