Rename "commit id" and "parent commit ids" to "id" and "parent ids".
In an effort to generalize terminology, the metadata keys "commit id"
and "parent commit ids" have had the "commit" term removed. The ID is
now the ID of the change, which may map to a commit, or it may not.

These sections of the spec were fleshed out to provide some examples of
how these fields would map for Git, Mercurial, and Subversion, and when
the IDs might be omitted.

On top of this, this key is no longer listed as required. It's now
listed as recommended. This is because an ID doesn't always make sense
for a change in a DiffX file.

Testing Done:
Built the spec and checked the updated sections and diffs.

Reviewed at https://reviews.reviewboard.org/r/11731/