Clarify some metadata and update optional/required/recommended status.
This updates the definitions and examples for some of the metadata,
making it more clear how some are used and what values are expected.

Some of the required keys are now marked as recommended instead, since
it doesn't always make sense to include them. "Required" show now be
used sparingly.

The file `stats` metadata key is also now recommended, instead of merely
optional, as it's a very useful thing to include if available for
statistics information.

Testing Done:
Built the docs. Read through the new changes. Checked the links for
errors.

Reviewed at https://reviews.reviewboard.org/r/11732/