Add initial documentation for pydiffx.
This provides some documentation on what pydiffx is for, how to install
it, an initial tutorial on writing a DiffX file using it, and a code
reference.

Upcoming changes will flesh out the tutorials a bit more, but this will
be sufficient for now, as we're getting close to the 1.0 release.

As part of this, we now require beanbag-docutils in order to build the
docs.

Testing Done:
Built the docs and read through them. Didn't see any immediate errors,
and tried to check spelling errors, but there's likely something that
slipped through.

Reviewed at https://reviews.reviewboard.org/r/11739/