Update the theme and Last Updated timestamp for the spec.
This switches the spec to use the Furo theme, which is a clean,
mobile-friendly theme with a lot of clean navigation and presentation
benefits (such as dark mode).

The Pygments theme has been switched to Autumn, which is a good light
theme that has suitable styling for diffs (many Pygments themes don't)
and fits in with Furo.

A Sphinx plugin was added that adds a handy Copy button for code
snippets, which can be nice for diffs and sample code.

And finally, the Last Updated timestamp for the spec is now based on the
modification time of the `*.rst` files in the `spec/` directory. This
helps keep it from being bumped up when docs for pydiffx or some other
part of the docs codebase is changed.

Testing Done:
Tested on desktop and mobile modes.

Went through all code snippets to make sure they looked reasonable.

Went through every page in the docs to check for formatting.

Reviewed at https://reviews.reviewboard.org/r/11741/