Add a utility function for returning encoded newlines for a type.
This introduces `get_newline_for_type()`, a utility function that takes
a `LineEndings` type value and returns newlines encoded as either ASCII
or the given encoding (with BOM stripped).

This logic was previously contained in `DiffXReader`, but is generally
useful, so it's being split out. That simplifies the `DiffXReader` code
as well, making some of the logic a lot more clear.

Testing Done:
Unit tests pass on Python 2 and 3.

Reviewed at https://reviews.reviewboard.org/r/11742/