Add a DOM function for generating the "stats" metadata.
This introduces `DiffX.generate_stats()`,
`DiffXChangeSection.generate_stats()`, and
`DiffXFileSection.generate_stats()`. These are used to generate metadata
containing useful statistics on the diff, such as the number of files,
changes, inserts/deletes, and changed lines in a diff.

Callers can call this immediately prior to writing the DiffX to a
stream, in order to provide this information for consumers of the diff.

Testing Done:
Unit tests pass on Python 2 and 3.

Reviewed at https://reviews.reviewboard.org/r/11743/