Add READMEs and a LICENSE file for the repository and pydiffx.
This adds some READMEs for both the main repository (for use on GitHub)
and for pydiffx (for use in the package) that describe DiffX as a whole
and pydiffx, respectively.

The pydiffx version is a shortened version of the main pydiffx
documentation, and links to the suitable spots in the main docs.

A LICENSE file was added to pydiffx and to the root of the repository.
This is the MIT license.

Testing Done:
Viewed both READMEs through a viewer emulating GitHub. Tested each link.

Reviewed at https://reviews.reviewboard.org/r/11750/