Clean up some text of the sections part of the spec and add more anchors.
This is a cleanup that goes through the various sections and cleans up
some spelling errors and ReST syntax issues, improves a few
descriptions, and adds anchors to all metadata.

Regarding the syntax errors, the main section tree at the top of the
page had parts that were indented too far, causing subsequent levels of
the tree to appear as quoted text and not sub-lists. This became more
apparent with the new theme being used for the documentation.

Testing Done:
Scanned the generated docs, making sure that the content rendered
correctly. Checked some of the new links to make sure they linked to
the right spots.

Reviewed at https://reviews.reviewboard.org/r/11753/