Fix a crash in DiffXFileSection.generate_stats with no diff.
`DiffXFileSection.generate_stats()` was assuming a diff was set, but
this isn't a requirement for this section (as there may just be metadata
changes).

This will now bail immediately if a diff is not set. Statistics in this
case won't be set.

Testing Done:
Unit tests pass.

Reviewed at https://reviews.reviewboard.org/r/11762/