Add six as a development dependency.
While `six` is a package-level dependency, it's really important to
install it before handling package operations. This is because
`setup.py` needs to import `get_package_version()` from `pydiffx`, which
eventually ends up importing `six` itself. To help mitigate this, we now
list `six` in `dev-requirements.txt`.

It's a trivial fix for the problem, but we may want to address this
another way down the road by avoiding the utility imports in `pydiffx`
if `six` is missing.