Fix setting diff types and avoid generating stats for binary diffs.
When setting a file as binary using the DOM, writing it out would fail,
due to the wrong parameter being passed to `DiffXWriter.write_diff()`.
This needed to be `diff_type=`, but it was being passed as `type=`. We
now use the option remapping support in `DiffXDOMWriter` to pass the
correct argument.

Statistics were also being generated for binary diffs, which didn't make
any sense. We now skip generating stats for these.

Testing Done:
Unit tests pass on Python 2 and 3.

Verified this fix in production.

Reviewed at https://reviews.reviewboard.org/r/11771/