Update the spec to include the line_endings option and trailing newline rules.
This introduces a new content option, `line_endings`, which is used to
define the explicit type of line endings the content uses. This is
primarily important for diff content, due to real-world encounters with
diffs that only appear to be in one form of line ending or another.

Along with this, information on content sections now explicitly state
that the section must end in a newline in the section's encoding.

Testing Done:
Built the docs and read through them. Checked that term links were
worked correctly. Checked for spelling errors.

Reviewed at https://reviews.reviewboard.org/r/11768/