Bump pydiffx to 2.0 dev and update Python compatibility.
This is the beginning of pydiffx 2.0, which will be largely compatible
with pydiffx 1.0 but supporting only modern versions of Python.

Support for Python 2.7 and 3.6 will be dropped. 3.10 and 3.11 will be
added.

This will eventually remove a dependency on `six` and add type
annotations, which will be very helpful as much of `pydiffx` involves
properly working with correct string types and dictionaries of
information.

The 1.x branch will still be maintained as needed, but new capabilities
will likely be placed only on 2.x.

Testing Done:
Successfully built packages.

Reviewed at https://reviews.reviewboard.org/r/12559/