Remove six from pydiffx.
This removes the dependency on `six` from the codebase. Now that this is
fully on Python 3, we no longer need to depend on this.

Our usage was pretty minimal. This updates everything to use raw types
and native functions.

Testing Done:
Unit tests passed.

Reviewed at https://reviews.reviewboard.org/r/12560/